eureka it works

+++ ArduinoCode/ArduinoController.h
@@ -6,19 +6,19 @@
 class ArduinoController {
   private:
     void parseData(String data);
-    void sendDataBack(int currentMotorPos[], int currentServoPos[]);
+    void sendDataBack(float currentMotorPos[], float currentServoPos[]);
 
   public:
     ArduinoController();
     void setup();
-    void loop(int newMotorPos[], int newServoPos[]);
-    void updateCurrentMSVals(int newMotorPos[], int newServoPos[]);
+    void loop(float newMotorPos[], float newServoPos[]);
+    void updateCurrentMSVals(float newMotorPos[], float newServoPos[]);
     static const int numMotors = 4; // Number of motors
     static const int numServos = 4; // Number of servos
-    int motorPositions[numMotors];
-    int servoPositions[numServos];
-    int motorPosCurr[numMotors];
-    int servoPosCurr[numServos];
+    float motorPositions[numMotors];
+    float servoPositions[numServos];
+    float motorPosCurr[numMotors];
+    float servoPosCurr[numServos];
 };
 
 #endif
